requirements.txt
requests
beautifulsoup4
pyyaml
yt-dlp
chromadb
numpy
ctranslate2>=4.0,<5
torchcodec==0.7.0
